Add migration for arbitrage outputs
